Cleanup menu collapse code
