+++ app/routes.ts
@@ -7,7 +7,7 @@ const routes: RouteConfig = [
   route("/color-scheme", "actions/color-scheme.ts"),
 
   route("", "pages/docs-layout.tsx", { id: "docs" }, [
-    route("/menu-collapse", "actions/menu-collapse.ts"),
+    route("/menu-collapse", "actions/menu-collapse/route.ts"),
     route("home", "pages/doc.tsx", { id: "home" }),
     route("changelog", "pages/doc.tsx", { id: "changelog" }),
     route("*", "pages/doc.tsx"),
